refactor(file): 컨트롤러 병합
